+++ api/src/prescription/campaign-participation/domain/models/CompetenceResult.js
