+++ api/tests/tooling/domain-builder/factory/build-competence-result.js
@@ -1,4 +1,4 @@
-import { CompetenceResult } from '../../../../src/shared/domain/models/CompetenceResult.js';
+import { CompetenceResult } from '../../../../src/prescription/campaign-participation/domain/models/CompetenceResult.js';
 
 const buildCompetenceResult = function ({
   id = 1,
